Added setup bash script.
Added a script that is copied over to /usr/local/sbin/ in the
container. This script is executed as last command and fires up the
Nginx webserver. This script checks if the client has passed in any
optional parameter. Depending on the data received, the Nginx
configuration file is modified on the fly through sed.