The Nginx's configuration now accepts parameter's from the client.
The Nginx's configuration had many parameters replaced by specific
keywords. There are replaced on the fly by an ad hoc Bash script
that is executed when the container is fired up. This script checks
if the client has passed in optional parameters. Finally, it
replaces those keywords with the proper values, thus creating an ad
hoc configuration.